stages/04-backend-build/output/etl/requirements.txt
supabase>=2.7.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pdfkit>=1.0.0
resend>=2.3.0
pandas>=2.2.0
pytest>=8.0.0
